requirements.txt
chromadb
faiss-cpu
langchain_google_genai
streamlit
PyPDF2
langchain
google-generativeai
python-dotenv
pytesseract
opencv-python
speechrecognition
pyaudio
transformers
torch
spacy
scikit-learn
whisper
pyttsx3
camelot-py
pandas
